refactor: consolidate test helpers into conftest.py
- Add make_ve_initialized_git_repo() and setup_task_directory() to
  tests/conftest.py as shared helpers
- Update 10 test files to import helpers from conftest instead of
  duplicating the code
- Remove redundant ArtifactIndex import in list_task_narratives()

This removes ~500 lines of duplicated test helper code.
